refactor tests and update readme
