+++ Nbuild/ResourceHelper.cs
@@ -5,10 +5,11 @@ namespace Nbuild
 {
     public class ResourceHelper
     {
-        public static void ExtractEmbeddedResource(string outputDir, string resourceLocation, string fileName)
+        public static void ExtractEmbeddedResource(string resourceLocation, string fileName)
         {
+
             using Stream stream = Assembly.GetExecutingAssembly().GetManifestResourceStream(resourceLocation) ?? throw new ArgumentException("No such resource", nameof(resourceLocation));
-            using var fileStream = new FileStream(Path.Combine(outputDir, fileName), FileMode.Create);
+            using var fileStream = new FileStream(fileName, FileMode.Create);
             for (int i = 0; i < stream.Length; i++)
             {
                 fileStream.WriteByte((byte)stream.ReadByte());
@@ -18,7 +19,7 @@ namespace Nbuild
 
         public static void ExtractEmbeddedResourceFromAssembly(string assembly, string resourceLocation, string fileName)
         {
-            using Stream stream = Assembly.LoadFile(assembly).GetManifestResourceStream(resourceLocation) ?? throw new ArgumentException("No such resource", nameof(resourceLocation));
+            using Stream stream = Assembly.Load(assembly).GetManifestResourceStream(resourceLocation) ?? throw new ArgumentException("No such resource", nameof(resourceLocation));
             using var fileStream = new FileStream(fileName, FileMode.Create);
             for (int i = 0; i < stream.Length; i++)
             {
