+++ nBackup/NBackup.cs
@@ -121,12 +121,6 @@ namespace nbackup
                     Console.WriteLine($"An exception occurred: {ex.Message}");
                 }
             }
-            else if (!string.IsNullOrEmpty(options.Source) &&
-                !string.IsNullOrEmpty(options.Destination) &&
-                !string.IsNullOrEmpty(options.Backup))
-            {
-                result = NBackup.Perform(options.Source, options.Destination, options.Backup);
-            }
             else
             {
                 result.Code = ResultHelper.FileNotFound;
