+++ nbuild.bat
@@ -1,11 +1,10 @@
 @echo off
 @REM Reference: https://learn.microsoft.com/en-us/visualstudio/msbuild/walkthrough-using-msbuild?view=vs-2022
 @REM This script is used to build the project using MSBuild
-@REM This file is generated by nbuild.exe
+@REM This file is generated by Nbuild.exe (Do not modify)
 @REM for more information, please visit <TBD>
 set BUILD_LOG=nbuild.log
 set PROJECT=nbuild.targets
-set TAG=tag.txt
 
 if "%1"=="" (
     set CMD=%PROJECT% -fl -flp:logfile=%BUILD_LOG%;verbosity=normal
