+++ ntools.sln
@@ -4,9 +4,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.0.31825.309
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "launcherTests", "LauncherTests\launcherTests.csproj", "{BE437ED5-EBDE-4B6C-9B4F-BA7E0D5898CC}"
-	ProjectSection(ProjectDependencies) = postProject
-		{4F511345-A840-4EC0-A547-F2EAA53935C8} = {4F511345-A840-4EC0-A547-F2EAA53935C8}
-	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".github", ".github", "{80B7EE8D-BFD5-4F31-BC37-103FDDACEA23}"
 EndProject
@@ -37,7 +34,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nbuild", "nbuild\Nbuild.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NbuildTests", "NbuildTests\NbuildTests.csproj", "{EFB85C8C-26C3-4A6F-919F-3534CD238975}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ngit", "Ngit\Ngit.csproj", "{4F511345-A840-4EC0-A547-F2EAA53935C8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LauncherTest", "LauncherTest\LauncherTest.csproj", "{EF541BB6-416C-4DDC-ACB5-B0BFECD2E9F4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -73,10 +70,10 @@ Global
 		{EFB85C8C-26C3-4A6F-919F-3534CD238975}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EFB85C8C-26C3-4A6F-919F-3534CD238975}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EFB85C8C-26C3-4A6F-919F-3534CD238975}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4F511345-A840-4EC0-A547-F2EAA53935C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4F511345-A840-4EC0-A547-F2EAA53935C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4F511345-A840-4EC0-A547-F2EAA53935C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4F511345-A840-4EC0-A547-F2EAA53935C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF541BB6-416C-4DDC-ACB5-B0BFECD2E9F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF541BB6-416C-4DDC-ACB5-B0BFECD2E9F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF541BB6-416C-4DDC-ACB5-B0BFECD2E9F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF541BB6-416C-4DDC-ACB5-B0BFECD2E9F4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
